changes rules corner
