+++ main.rb
@@ -2,6 +2,7 @@ require_relative 'Game_of_life'
 require_relative 'rules'
 require_relative 'death'
 require_relative 'frame_rules'
+require_relative 'corner_rules'
 game = Game_Life.new
 game.start
