+++ rules.rb
@@ -11,19 +11,19 @@ class Rules < Game_Life
     end
 		# left corner  top
     if (row == 0 && col == 0) 
-      aliveCells = left_corner_top(row,col)
+      aliveCells = Corner_Rules.new.left_corner_top(row,col)
     end
     #rigth corner top
     if (row == 0 && col == @@board_size - 1) 
-      aliveCells = rigth_corner_top(row,col)
+      aliveCells = Corner_Rules.new.rigth_corner_top(row,col)
     end
     #left corner down
     if (row == @@board_size - 1 && col == 0) 
-      aliveCells = left_corner_down(row,col)
+      aliveCells = Corner_Rules.new.left_corner_down(row,col)
     end
     #rigth corner down
     if (row == @@board_size - 1 && col == @@board_size - 1) 
-      aliveCells = rigth_corner_down(row,col)        
+      aliveCells = Corner_Rules.new.rigth_corner_down(row,col)        
     end
     #edge central top
     if (row == 0 && col != 0 && col != @@board_size - 1) 
@@ -74,136 +74,6 @@ class Rules < Game_Life
     end
     if (@@board[row + 1][col + 1] == 1) 
          aliveCells += 1 
-    end
-    return aliveCells
-  end
-
-  def left_corner_top(row,col)
-    aliveCells = 0
-    #top
-    if (@@board[@@board_size - 1][@@board_size - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[@@board_size - 1][0] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[@@board_size - 1][1] == 1) 
-        aliveCells+=1
-    end
-    #same row 
-    if (@@board[row][@@board_size - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row][col + 1] == 1) 
-        aliveCells+=1
-    end
-    #under
-    if (@@board[row + 1][@@board_size - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row + 1][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row + 1][col + 1] == 1) 
-        aliveCells+=1
-    end
-    return aliveCells
-  end
-
-  def rigth_corner_top(row,col)
-    aliveCells = 0
-    #top
-    if (@@board[@@board_size - 1][col - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[@@board_size - 1][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[@@board_size - 1][0] == 1) 
-        aliveCells+=1
-    end
-    #same row 
-    if (@@board[row][col - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row][0] == 1) 
-        aliveCells+=1
-    end
-    #under
-    if (@@board[row + 1][col - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row + 1][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row + 1][0] == 1) 
-        aliveCells+=1
-    end
-     # puts "Celulas vivas rigth corner top"
-     # puts aliveCells
-     return aliveCells
-  end
-
-  def left_corner_down(row,col)
-    aliveCells = 0
-    #top
-    if (@@board[row - 1][@@board_size - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row - 1][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row - 1][col + 1] == 1) 
-        aliveCells+=1
-    end
-    #same row 
-    if (@@board[row][@@board_size - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row][col + 1] == 1) 
-        aliveCells+=1
-    end
-    #under
-    if (@@board[0][@@board_size - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[0][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[0][col + 1] == 1) 
-        aliveCells+=1
-    end
-    return aliveCells
-  end
-
-  def rigth_corner_down(row,col)
-    aliveCells = 0
-    #top
-    if (@@board[row - 1][col - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row - 1][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row - 1][0] == 1) 
-        aliveCells+=1
-    end
-    #same row 
-    if (@@board[row][col - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[row][0] == 1) 
-        aliveCells+=1
-    end
-    #under
-    if (@@board[0][col - 1] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[0][col] == 1) 
-        aliveCells+=1
-    end
-    if (@@board[0][0] == 1) 
-        aliveCells+=1
     end
     return aliveCells
   end
